Introducing SYSTEM_USER
From now on, there will be one special user called system, that will not be able to login, will not show up as registered user, but it will be possible to assign it to specific actions.
Its user ID should be 0, but can be anything as it's represented by static constant present on WebAuthManager class and is called SYSTEM_USER_ID. All account actions are blocked for the system account.
It's still possible to access system user by methods like getUser, userExists and resolveUserKey by specifying additional parameter hideSystemUser and setting it to false. System user will never by returned within collection of users.
If there is an action, that can result in the system object being assigned to or owning an object, all calls to getUser, userExists and resolveUserKey from within code related to that functionality should be made with the hideSystemUser parameter changed accordingly.