Improvements to web module caching
Fixed problems with some cached files being obsolete and resulting in page that can't be rendered.
Not all chunks have hash embedded so if they are changed.

Also removed version from the main chunks. Modules like luxon will rarely be changed, so they can be stored in cache for a long time, version in name would prevent that..